Refactor some things to make an ff_mocks module that can be used outside this repo.
