+++ dcicutils/qa_utils.py
@@ -1339,62 +1339,3 @@ def make_mock_beanstalk_environment_variables(var_str):
     spec2 = [{"Namespace": _NAMESPACE_ENVIRONMENT_VARIABLE, "OptionName": name, "Value": value}
              for name, value in [spec.split("=") for spec in var_str.split(",")]]
     return spec0 + spec1 + spec2
-
-
-_MOCK_APPLICATION_NAME = "4dn-web"
-_MOCK_SERVICE_USERNAME = 'service-accout-name'
-_MOCK_SERVICE_PASSWORD = 'service-accout-pw'
-_MOCK_APPLICATION_OPTIONS_PARTIAL = (
-    f"MOCK_USERNAME={_MOCK_SERVICE_USERNAME},MOCK_PASSWORD={_MOCK_SERVICE_PASSWORD},ENV_NAME="
-)
-
-
-class MockBoto4DNLegacyElasticBeanstalkClient(MockBotoElasticBeanstalkClient):
-
-    DEFAULT_MOCKED_BEANSTALKS = [
-        make_mock_beanstalk("fourfront-cgapdev"),
-        make_mock_beanstalk("fourfront-cgapwolf"),
-        make_mock_beanstalk("fourfront-cgap"),
-        make_mock_beanstalk("fourfront-cgaptest"),
-        make_mock_beanstalk("fourfront-webdev"),
-        make_mock_beanstalk("fourfront-hotseat"),
-        make_mock_beanstalk("fourfront-mastertest"),
-        make_mock_beanstalk("fourfront-green", cname="fourfront-green.us-east-1.elasticbeanstalk.com"),
-        make_mock_beanstalk("fourfront-blue"),
-    ]
-
-    MOCK_APPLICATION_NAME = _MOCK_APPLICATION_NAME
-    MOCK_SERVICE_USERNAME = _MOCK_SERVICE_USERNAME
-    MOCK_SERVICE_PASSWORD = _MOCK_SERVICE_PASSWORD
-
-    DEFAULT_MOCKED_CONFIGURATION_SETTINGS = [
-        {
-            "ApplicationName": _MOCK_APPLICATION_NAME,
-            "EnvironmentName": env_name,
-            "DeploymentStatus": "deployed",
-            "OptionSettings": make_mock_beanstalk_environment_variables(_MOCK_APPLICATION_OPTIONS_PARTIAL + env_name),
-        }
-        for env_name in [beanstalk["EnvironmentName"] for beanstalk in DEFAULT_MOCKED_BEANSTALKS]
-    ]
-
-    @classmethod
-    def all_legacy_beanstalk_names(cls):
-        return [beanstalk["EnvironmentName"] for beanstalk in cls.DEFAULT_MOCKED_BEANSTALKS]
-
-
-class MockBotoFooBarElasticBeanstalkClient(MockBotoElasticBeanstalkClient):
-
-    DEFAULT_MOCKED_BEANSTALKS = [
-        make_mock_beanstalk("fourfront-foo"),
-        make_mock_beanstalk("fourfront-bar"),
-    ]
-
-    DEFAULT_MOCKED_CONFIGURATION_SETTINGS = [
-        {
-            "ApplicationName": _MOCK_APPLICATION_NAME,
-            "EnvironmentName": env_name,
-            "DeploymentStatus": "deployed",
-            "OptionSettings": make_mock_beanstalk_environment_variables(_MOCK_APPLICATION_OPTIONS_PARTIAL + env_name),
-        }
-        for env_name in [beanstalk["EnvironmentName"] for beanstalk in DEFAULT_MOCKED_BEANSTALKS]
-    ]
