+++ test/test_ff_utils.py
@@ -8,11 +8,11 @@ import shutil
 import time
 
 from botocore.exceptions import ClientError
-from dcicutils import es_utils, ff_utils, s3_utils, beanstalk_utils
+from dcicutils import es_utils, ff_utils, s3_utils
 from dcicutils.misc_utils import make_counter, remove_prefix, remove_suffix, check_true
+from dcicutils.ff_mocks import mocked_s3utils, TestScenarios
 from dcicutils.qa_utils import (
-    check_duplicated_items_by_key, MockResponse, ignored, MockBoto3, MockBotoSQSClient, MockBotoS3Client,
-    override_environ, raises_regexp, MockBotoFooBarElasticBeanstalkClient, make_mock_beanstalk_cname
+    check_duplicated_items_by_key, ignored, raises_regexp, MockResponse, MockBoto3, MockBotoSQSClient,
 )
 from types import GeneratorType
 from unittest import mock
@@ -235,166 +235,48 @@ def test_unified_authentication_decoding(integrated_ff):
             assert 'Must provide a valid authorization key or ff' in str(exec_info.value)
 
 
-# Here we set up some variables, auxiliary functions, and mocks containing common values needed for testing
-# of the next several functions so that the functions don't have to set them up over
-# and over again in each test.
-
-some_server = 'http://localhost:8000'
-some_auth_key, some_auth_secret = some_auth_tuple = ('mykey', 'mysecret')
-some_badly_formed_auth_tuple = ('mykey', 'mysecret', 'other-junk')  # contains an extra element
-some_auth_dict = {'key': some_auth_key, 'secret': some_auth_secret}
-some_auth_dict_with_server = {'key': some_auth_key, 'secret': some_auth_secret, 'server': some_server}
-some_badly_formed_auth_dict = {'kee': some_auth_key, 'secret': some_auth_secret}  # 'key' is misspelled
-
-foo_env = 'fourfront-foo'
-foo_env_auth_key, foo_env_auth_secret = foo_env_auth_tuple = ('fookey', 'foosecret')
-foo_env_auth_dict = {
-    'key': foo_env_auth_key,
-    'secret': foo_env_auth_secret
-}
-
-FOURFRONT_FOO_HEALTH_PAGE = {
-    s3_utils.HealthPageKey.BEANSTALK_ENV: "fourfront-foo",
-    s3_utils.HealthPageKey.ELASTICSEARCH: "http://fourfront-foo.elasticsearch.whatever",
-}
-
-bar_env = 'fourfront-bar'
-bar_env_auth_key, bar_env_auth_secret = bar_env_auth_tuple = ('barkey', 'barsecret')
-bar_env_url = f"http://{make_mock_beanstalk_cname('fourfront-bar')}/"
-bar_env_url_trimmed = bar_env_url.rstrip('/')
-bar_env_auth_dict = {
-    'key': bar_env_auth_key,
-    'secret': bar_env_auth_secret,
-    'server': bar_env_url,
-}
-bar_env_default_auth_dict = {'default': bar_env_auth_dict}
-
-FOURFRONT_BAR_HEALTH_PAGE = {
-    s3_utils.HealthPageKey.BEANSTALK_ENV: "fourfront-bar",
-    s3_utils.HealthPageKey.ELASTICSEARCH: "http://fourfront-bar.elasticsearch.whatever",
-}
-
-SSE_CUSTOMER_KEY = 'SSECustomerKey'
-SSE_CUSTOMER_ALGORITHM = 'SSECustomerAlgorithm'
-
-
-def _access_key_home(bs_env):
-    """
-    Constructs the location in our S3 mock where the stashed access keys live for a given bs_env.
-
-    In our MockBotoS3, buckets and keys in s3 are represented as bucket/key in a MockFileSystem M
-    that is used by the S3 mock to hold its contents. So buckets are just the toplevel folders in M,
-    and keys are below that, and to preset a mock with contents for a given bucket and key, one just
-    creates bucket/key as a filename in M.  This function, _access_key_home will return the appropriate
-    filename for such a mock file system to contain the access keys for a given bs_env.
-    """
-    return os.path.join(s3_utils.s3Utils.SYS_BUCKET_TEMPLATE % bs_env, s3_utils.s3Utils.ACCESS_KEYS_S3_KEY)
-
-
-class MockBotoS3WithSSE(MockBotoS3Client):
-    """
-    This is a specialized mock for a boto3 S3 client that does server-side encryption (SSE).
-    We don't actually test that encryption is done, since that would be done by boto3, but we set up so that
-    that calls to any methods on the mock use certain required keyword arguments beyond the ordinary methods
-    that do the ordinary work of the method. e.g., for this mock definition:
-
-        def upload_file(self, Filename, Bucket, Key, **kwargs)
-            ...
-
-    if the call does not look like:
-
-        upload_File(Filename=..., Bucket=..., Key=...,
-                    SSECustomerKey='shazam', SSECustomerAlgorithm='AES256')
-
-    where the mock value 'shazam' is a wired value supplied in calls to the mock and 'AES256' is a constant we
-    expect is consistently used throughout this code base, then an error is raised in testing. The intent is to
-    make sure the calls are done with a certain degree of consistency.
-    """
-
-    SSE_ENCRYPT_KEY = 'shazam'
-
-    MOCK_REQUIRED_ARGUMENTS = {
-        # Our s3 mock for this test must check that all API calls pass these required arguments.
-        SSE_CUSTOMER_KEY: SSE_ENCRYPT_KEY,
-        SSE_CUSTOMER_ALGORITHM: "AES256"
-    }
-
-    MOCK_STATIC_FILES = {
-        # Our s3 mock presumes access keys are stashed in a well-known key
-        # in the system bucket corresponding to the given beanstalk.
-        _access_key_home(foo_env): json.dumps(foo_env_auth_dict),
-        _access_key_home(bar_env): json.dumps(bar_env_default_auth_dict)
-    }
-
-
 @contextlib.contextmanager
-def mocked_s3_with_sse():
-    """
-    This context manager sets up a mock version of boto3 for use by s3_utils and ff_utils during the context
-    of its test. It also sets up the S3_ENCRYPT_KEY environment variable with a sample value for testing,
-    and it sets up a set of mocked beanstalks for fourfront-foo and fourfront-bar, so that s3_utils will not
-    get confused when it does discovery operations to find them.
-    """
-    # First we make a mocked boto3 that will use an S3 mock with mock server side encryption.
-    mock_boto3 = MockBoto3(s3=MockBotoS3WithSSE, elasticbeanstalk=MockBotoFooBarElasticBeanstalkClient)
-    # Now we arrange that both s3_utils and ff_utils modules share the illusion that our mock IS the boto3 library
-    with mock.patch.object(s3_utils, "boto3", mock_boto3):
-        with mock.patch.object(ff_utils, "boto3", mock_boto3):
-            with mock.patch.object(beanstalk_utils, "boto3", mock_boto3):
-                with mock.patch.object(s3_utils.EnvManager, "fetch_health_page_json") as mock_fetch_health_page_json:
-
-                    # This is all that's needed for s3Utils to initialize an EnvManager.
-                    # We might have to add more later.
-                    def mocked_fetch_health_page_json(url, use_urllib=True):
-                        ignored(use_urllib)  # we don't test this
-                        if 'fourfront-foo' in url:
-                            return FOURFRONT_FOO_HEALTH_PAGE
-                        elif 'fourfront-bar' in url:
-                            return FOURFRONT_BAR_HEALTH_PAGE
-                        else:
-                            raise NotImplementedError(f"Mock can't handle URL: {url}")
-
-                    mock_fetch_health_page_json.side_effect = mocked_fetch_health_page_json
-                    # The mocked encrypt key is expected by various tools in the s3_utils module to be supplied
-                    # as an environment variable (i.e., in os.environ), so this sets up that environment variable.
-                    with override_environ(S3_ENCRYPT_KEY=MockBotoS3WithSSE.SSE_ENCRYPT_KEY):
-                        yield
+def mocked_s3utils_with_sse():
+    with mocked_s3utils(beanstalks=['fourfront-foo', 'fourfront-bar'], require_sse=True):
+        yield
 
 
 def test_unified_authentication_unit():
 
-    with mocked_s3_with_sse():
+    ts = TestScenarios
+
+    with mocked_s3utils_with_sse():
 
         # When supplied a basic auth tuple, (key, secret), and a fourfront environment, the tuple is returned.
-        auth1 = ff_utils.unified_authentication(some_auth_tuple, foo_env)
-        assert auth1 == some_auth_tuple
+        auth1 = ff_utils.unified_authentication(ts.some_auth_tuple, ts.foo_env)
+        assert auth1 == ts.some_auth_tuple
 
         # When supplied a password and a fourfront environment,
-        auth2 = ff_utils.unified_authentication(some_auth_dict, foo_env)
-        assert auth2 == some_auth_tuple  # Given an auth dict, the result is canonicalized to an auth tuple
+        auth2 = ff_utils.unified_authentication(ts.some_auth_dict, ts.foo_env)
+        assert auth2 == ts.some_auth_tuple  # Given an auth dict, the result is canonicalized to an auth tuple
 
         # A deprecated form of the dictionary has an extra layer of wrapper we have to strip off.
         # Hopefully calls like this are being rewritten, but for now we continue to support the behavior.
-        auth3 = ff_utils.unified_authentication({'default': some_auth_dict}, foo_env)
-        assert auth3 == some_auth_tuple  # Given an auth dict, the result is canonicalized to an auth tuple
+        auth3 = ff_utils.unified_authentication({'default': ts.some_auth_dict}, ts.foo_env)
+        assert auth3 == ts.some_auth_tuple  # Given an auth dict, the result is canonicalized to an auth tuple
 
         # This tests that both forms of default credentials are supported remotely.
 
-        auth4 = ff_utils.unified_authentication(None, foo_env)
+        auth4 = ff_utils.unified_authentication(None, ts.foo_env)
         # Check that the auth dict fetched from server is canonicalized to an auth tuple...
-        assert auth4 == foo_env_auth_tuple
+        assert auth4 == ts.foo_env_auth_tuple
 
-        auth5 = ff_utils.unified_authentication(None, bar_env)
+        auth5 = ff_utils.unified_authentication(None, ts.bar_env)
         # Check that the auth dict fetched from server is canonicalized to an auth tuple...
-        assert auth5 == bar_env_auth_tuple
+        assert auth5 == ts.bar_env_auth_tuple
 
         with raises_regexp(ValueError, "Must provide a valid authorization key or ff environment."):
             # The .unified_authentication operation checks that an auth given as a tuple has length 2.
-            ff_utils.unified_authentication(some_badly_formed_auth_tuple, foo_env)
+            ff_utils.unified_authentication(ts.some_badly_formed_auth_tuple, ts.foo_env)
 
         with raises_regexp(ValueError, "Must provide a valid authorization key or ff environment."):
             # The .unified_authentication operation checks that an auth given as a dict has keys 'key' and 'secret'.
-            ff_utils.unified_authentication(some_badly_formed_auth_dict, foo_env)
+            ff_utils.unified_authentication(ts.some_badly_formed_auth_dict, ts.foo_env)
 
         with raises_regexp(ValueError, "Must provide a valid authorization key or ff environment."):
             # If the first arg (auth) is None, the second arg (ff_env) must not be.
@@ -449,21 +331,23 @@ def test_unified_authentication_prod_envs_integrated_only():
 
 def test_get_authentication_with_server_unit():
 
-    with mocked_s3_with_sse():
+    ts = TestScenarios
 
-        key1 = ff_utils.get_authentication_with_server(bar_env_auth_dict, None)
+    with mocked_s3utils_with_sse():
+
+        key1 = ff_utils.get_authentication_with_server(ts.bar_env_auth_dict, None)
         assert isinstance(key1, dict)
         assert {'server', 'key', 'secret'} <= set(key1.keys())
-        assert key1['key'] == bar_env_auth_key
-        assert key1['secret'] == bar_env_auth_secret
-        assert key1['server'] == bar_env_url_trimmed
+        assert key1['key'] == ts.bar_env_auth_key
+        assert key1['secret'] == ts.bar_env_auth_secret
+        assert key1['server'] == ts.bar_env_url_trimmed
 
-        key2 = ff_utils.get_authentication_with_server(bar_env_default_auth_dict, None)
+        key2 = ff_utils.get_authentication_with_server(ts.bar_env_default_auth_dict, None)
         assert isinstance(key2, dict)
         assert {'server', 'key', 'secret'} <= set(key2.keys())
         assert key2 == key1
 
-        key3 = ff_utils.get_authentication_with_server(None, bar_env)
+        key3 = ff_utils.get_authentication_with_server(None, ts.bar_env)
         assert isinstance(key2, dict)
         assert {'server', 'key', 'secret'} <= set(key3.keys())
         assert key3 == key1
@@ -471,7 +355,7 @@ def test_get_authentication_with_server_unit():
         with raises_regexp(ValueError, 'ERROR GETTING SERVER'):
             # The mocked foo_env, unlike the bar_env, is missing a 'server' key in its dict,
             # so it will be rejected as bad data by .get_authentication_with_server().
-            ff_utils.get_authentication_with_server(foo_env_auth_dict, None)
+            ff_utils.get_authentication_with_server(ts.foo_env_auth_dict, None)
 
         with raises_regexp(ValueError, 'ERROR GETTING SERVER'):
             # If the auth is not provided (or None), the ff_env must be given (and not None).
@@ -591,6 +475,8 @@ def test_authorized_request_integrated(integrated_ff):
 
 def test_get_metadata_unit():
 
+    ts = TestScenarios
+
     # The first part of this function sets up some common tools and then we test various scenarios
 
     counter = make_counter()  # used to generate some sample data in mock calls
@@ -623,13 +509,13 @@ def test_get_metadata_unit():
         It returns mock data that identifies itself as what was asked for.
         """
         ignored(ff_env, kwargs)
-        assert url.startswith(bar_env_url)
+        assert url.startswith(ts.bar_env_url)
         assert retry_fxn == unsupplied, "The retry_fxn argument was not expected by this mock."
         assert verb == 'GET'
-        assert auth == bar_env_auth_dict
+        assert auth == ts.bar_env_auth_dict
         return MockResponse(json={
             'mock_data': counter(),  # just so we can tell calls apart
-            '@id': remove_prefix(bar_env_url_trimmed, remove_suffix("?datastore=database", url))
+            '@id': remove_prefix(ts.bar_env_url_trimmed, remove_suffix("?datastore=database", url))
         })
 
     def make_mocked_authorized_request_erring(status_code, json):
@@ -657,7 +543,7 @@ def test_get_metadata_unit():
             # that datum is returned correctly.
             mock_authorized_request.side_effect = mocked_authorized_request
             # Do the call, which will bottom out at our mock call
-            res_w_key = ff_utils.get_metadata(test_item, key=bar_env_auth_dict, check_queue=check_queue, **kwargs)
+            res_w_key = ff_utils.get_metadata(test_item, key=ts.bar_env_auth_dict, check_queue=check_queue, **kwargs)
             # Check that the data flow back from our mock authorized_request call did what we expect
             assert res_w_key == {'@id': test_item_id, 'mock_data': n}
             # Check that the call out to the mock authorized_request is the thing we think.
@@ -668,8 +554,8 @@ def test_get_metadata_unit():
             # (d) on certain calls, an additional query string (e.g., "?datastore=database") will be used in the
             #     function we're testing in some cases (e.g., because check_queue=True is used), so we take that suffix
             #     as an argument and test for it.
-            mock_authorized_request.assert_called_with(bar_env_url.rstrip('/') + test_item_id + expect_suffix,
-                                                       auth=bar_env_auth_dict,
+            mock_authorized_request.assert_called_with(ts.bar_env_url.rstrip('/') + test_item_id + expect_suffix,
+                                                       auth=ts.bar_env_auth_dict,
                                                        verb='GET')
 
         with mock.patch.object(ff_utils, "stuff_in_queues", make_mocked_stuff_in_queues()):
@@ -678,12 +564,12 @@ def test_get_metadata_unit():
         with mock.patch.object(ff_utils, "stuff_in_queues", make_mocked_stuff_in_queues(return_value=True)):
             with raises_regexp(ValueError, "environment name"):
                 test_it(1, check_queue=True, expect_suffix="?datastore=database")
-            test_it(1, check_queue=True, expect_suffix="?datastore=database", ff_env=bar_env)
+            test_it(1, check_queue=True, expect_suffix="?datastore=database", ff_env=ts.bar_env)
 
         with mock.patch.object(ff_utils, "stuff_in_queues", make_mocked_stuff_in_queues(return_value=False)):
             with raises_regexp(ValueError, "environment name"):
                 test_it(2, check_queue=True)
-            test_it(2, check_queue=True, ff_env=bar_env)
+            test_it(2, check_queue=True, ff_env=ts.bar_env)
 
     # Now we test the error scenarios...
 
@@ -695,7 +581,7 @@ def test_get_metadata_unit():
         #            as if it were what was requested. See next check.
         mock_authorized_request.side_effect = make_mocked_authorized_request_erring(500, json=None)
         with raises_regexp(Exception, "Cannot get json"):
-            ff_utils.get_metadata(test_item, key=bar_env_auth_dict, check_queue=False)
+            ff_utils.get_metadata(test_item, key=ts.bar_env_auth_dict, check_queue=False)
 
     # This tests that the error message JSON can be obtained if the body is well-formed.
 
@@ -705,7 +591,7 @@ def test_get_metadata_unit():
         # TODO: Consider whether get_metadata() would be better off doing a .raise_For_status() -kmp 25-Oct-2020
         error_message_json = {'message': 'foo'}
         mock_authorized_request.side_effect = make_mocked_authorized_request_erring(500, json=error_message_json)
-        assert ff_utils.get_metadata(test_item, key=bar_env_auth_dict, check_queue=False) == error_message_json
+        assert ff_utils.get_metadata(test_item, key=ts.bar_env_auth_dict, check_queue=False) == error_message_json
 
 
 @pytest.mark.integratedx
@@ -1740,21 +1626,22 @@ SAMPLE_COUNTS_MATCH = {
 def test_are_counts_even_unit(expect_match, sample_counts):
 
     unsupplied = object()
+    ts = TestScenarios
 
     def mocked_authorized_request(url, auth=None, ff_env=None, verb='GET',
                                   retry_fxn=unsupplied, **kwargs):
         print("URL=%s auth=%s ff_env=%s verb=%s" % (url, auth, ff_env, verb))
         ignored(ff_env, kwargs)
-        assert auth == bar_env_auth_dict
+        assert auth == ts.bar_env_auth_dict
         assert verb == 'GET'
         assert retry_fxn == unsupplied
 
         return MockResponse(json=sample_counts)
 
-    with mocked_s3_with_sse():
+    with mocked_s3utils_with_sse():
         with mock.patch.object(ff_utils, "authorized_request", mocked_authorized_request):
             with mock.patch.object(s3_utils.s3Utils, "get_access_keys",
-                                   return_value=bar_env_auth_dict):
+                                   return_value=ts.bar_env_auth_dict):
 
                 counts_are_even, totals = ff_utils.get_counts_summary(env='fourfront-foo')
                 print("expect_match=", expect_match, "counts_are_even=", counts_are_even, "totals=", totals)
